requirements.txt
Django==1.6.5
Pillow
sorl-thumbnail